Update methods to maintain convention
By convention, methods that end in a question mark return a boolean.
To adhere to this, the value array is indexed such that the higher
scoring poker hands are closer to 0 than the lower scording hands. This
allows us to find the best score in a hand by returning the index of the
first true return from the Value class.

In the winner method we now return the hand that has the lowest value.