Implement Straight
Almost the same implementation as StraightFlush except we check if all
of the suits are the same. If they are the same, we return early because
we might have a StraightFlush (all of the suits are the same).

We then check if the cards in sequence by sorting and checking if
every card is delta 1 of each other.