Implement FullHouse
Implement FullHouse by first checking if the hand has a three of a kind
by utilizing ThreeOfAKind. Then check if there is a rank in the hand
that only has 2 occurences.

The difference between this way of checking for a 'pair' and our
implemented Pair class is that the Pair class checks if there are any
pairs at all, thus a hand that contains a 3 of a kind would return
true if passed to the Pair class. In our implementation of FullHouse
we need to make sure that the three of a kind in the hand don't
interfere with checking for a pair.