Refactor Straights and Flushes
Change StraightFlush to utilize the Straight and Flush service objects.
This makes the StraightFlush implementation trivial to implement.

Also change Flush and Straight to more accurately return when given
a flush or straight as Flush no longer checks if the hand is not in a
sequence and Straight no longer checks if the hand is not all the same
suit.