Implement TwoPair tie scenario
When both poker hands are a TwoPair, using Comparable, the TwoPair
entity will return the winner. If both hands are exactly alike, 'tie' is
returned.
Very open to syntax suggestions here to make this easier to read.

Very open to syntax suggestions in find_two_pair.rb to make it easier
to read.