Add return condition to FindFullHouse
For the FindThreeOfAKind service object and the ThreeOfAKind entity
to work, a second return early condition needed to be added to
FindFullHouse as poker hands consisting of 'three of a kind' were
getting recognized as a 'full house'.

Now FindFullHouse return early if there is no pair in the hand, not
including the 'three of a kind' in the hand.

A proper implementation of FindFullHouse and FullHouse is coming.