Implement ThreeOfAKind tie scenario
When both poker hands are a ThreeOfAKind, using Comparable, the
ThreeOfAKind entity will return the winner. If both hands are exactly
alike, 'tie' is returned.
Very open to syntax suggestions here to make this easier to read.

Very open to syntax suggestions in find_three_of_a_kind.rb to make it
easier to read.