Add return condition to FindStraightFlush
For the FindStraightFlush service object and the StraightFlush entity
to work, a second return early condition needed to be added to
FindStraightFlush as poker hands consisting of a 'straight' were
getting recognized as a 'straight flush'.

Now FindStraghtFlush returns early if there is no straight in the hand.

A proper implementation of FindFullHouse and FullHouse is coming.