Implement StraightFlush tie scenario
When both poker hands are a StraightFlush we need to decide a winner.
We use the <=> operator to decide this because we can compare the
ranks of the StraightFlush hands element to element until one hand beats
the other.

The <=> comparison is not intended to be used between poker hands at
this point unless they are of the same Entity.