Implement RoyalFlush tie scenario
While there can never be a RoyalFlush tie scenario in a real game of
poker, there can be two royal flush's drawn in this version.

When both poker hands are a RoyalFlush we need to decide a winner.
In this scenario we use the <=> to return a tie when both poker hands
are a RoyalFlush because that is the only outcome of two RoyalFlush
hands.

The <=> comparison is not intended to be used between poker hands at
this point unless they are of the same Entity, to resolve ties.