Determine winner and print hands and winner
When we compare two poker hands we need to determine the winner. We
simply compare their strengths because the strength associated with
a hand represents the best poker hand within that hand. If the strengths
are the same we need to use the <=> operator to decide who wins in
a tie.

We print the two poker hands and the winner to the screen because we
want visual confirmation on the winner of the poker game.