Test FindFlush service object
We need to verify that we are getting an instance of the Flush entity
because FindFlush returns that entity, which is used to determine the
winner if a hand contains a flush.

We need to verify the :flush attribute on the Flush entity because we
cannot perform comparisons during a tie without properly saving the
Card objects provided when a flush is identified.

We need to verify that FindFlush only returns a Flush entity when a
flush is found because misidentifying a poker hand as a flush would
result in innacurate indentification of a winner.