Test FindHighCard service object
We need to verify that we are getting an instance of the HighCard entity
because FindHighCard returns that entity, which is used to determine the
winner if a hand contains no better poker hand.

We need to verify the :cards attribute on the HighCard entity because we
cannot perform comparisons during a tie without properly saving the
Card objects provided when a HighCard is identified.