Spec for FindPair service object
We need to verify that we are getting an instance of the Pair entity
because FindPair returns that entity, which is used to determine the
winner if a hand contains a pair.

We need to verify the :cards, :pair, :other_cards attributes on the Pair
entity because we cannot perform comparisons during a tie without
properly saving the Card objects provided when a pair is identified.

We need to verify that FindPair only returns a Pair entity when a
pair is found because misidentifying a poker hand as a pair would
result in inaccurate identification of a winner.