Spec for FindTwoPair service object
We need to verify that we are getting an instance of the TwoPair entity
because FindTwoPair returns that entity, which is used to determine the
winner if a hand contains a two pair.

We need to verify the :cards, :pairs, :other_cards attributes on the
TwoPair entity because we cannot perform comparisons during a tie
without properly saving the Card objects provided when a two pair is
identified.

We need to verify that FindTwoPair only returns a TwoPair entity when a
two pair is found because misidentifying a poker hand as a two pair
would result in inaccurate identification of a winner.