Add :cards attr to TwoPair entity
Separate :cards, :pairs, and :other_card such that they hold Card
objects.

If we want to test our poker hand entities we can't strip information
off of the Card objects in the array and overwrite them. We need this
information if we want to be able to properly test and perhaps use
this information in the future.

Instead we put the logic in the <=> method because that's only where it
is needed.