Spec for FindThreeOfAKind service object
We need to verify that we are getting an instance of the ThreeOfAKind
entity because FindThreeOfAKind returns that entity, which is used to
determine the winner if a hand contains a three of a kind.

We need to verify the :cards, :set, :other_cards attributes on the
ThreeOfAKind entity because we cannot perform comparisons during a tie
without properly saving the Card objects provided when a pair is
identified.

We need to verify that FindThreeOfAKind only returns a ThreeOfAKind
entity when a pair is found because misidentifying a poker hand as a
ThreeOfAKind would result in inaccurate identification of a winner.