Spec for FindStraight service object
We need to verify that we are getting an instance of the Straight entity
because FindStraight returns that entity, which is used to determine the
winner if a hand contains a straight.

We need to verify the :cards attribute on the Straight entity because we
cannot perform comparisons during a tie without properly saving the Card
objects provided when a straight is identified.

We need to verify that FindStraight only returns a Straight entity when
a pair is found because misidentifying a poker hand as a straight would
result in inaccurate identification of a winner.