Spec for FindFullHouse service object
We need to verify that we are getting an instance of the FullHouse
entity because FindFullHouse returns that entity, which is used to
determine the winner if a hand contains a full house.

We need to verify the :cards, :set, :pair attributes on the
FullHouse entity because we cannot perform comparisons during a tie
without properly saving the Card objects provided when a full house is
identified.

We need to verify that FindFullHouse only returns a FullHouse
entity when a full house is found because misidentifying a poker hand as
a FullHouse would result in inaccurate identification of a winner.