Spec for FindFourOfAKind service object
We need to verify that we are getting an instance of the FourOfAKind
entity because FindFourOfAKind returns that entity, which is used to
determine the winner if a hand contains a four of a kind.

We need to verify the :cards, :quads, :other_card attributes on the
FourOfAKind entity because we cannot perform comparisons during a tie
without properly saving the Card objects provided when a four of a kind
is identified.

We need to verify that FindFourOfAKind only returns a FourOfAKind
entity when a full house is found because misidentifying a poker hand as
a FourOfAKind would result in inaccurate identification of a winner.