Spec for FindStraightFlush service object
We need to verify that we are getting an instance of the StraightFlush
entity because FindStraightFlush returns that entity, which is used to
determine the winner if a hand contains a straight flush.

We need to verify the :cards, attributes on the StraightFlush entity
because we cannot perform comparisons during a tie without properly
saving the Card objects provided when a straight flush is identified.

We need to verify that FindStraightFlush only returns a StraightFlush
entity when a straight flush is found because misidentifying a poker
hand as a StraightFlush would result in inaccurate identification of a
winner.