Spec for FindRoyalFlush service object
We need to verify that we are getting an instance of the RoyalFlush
entity because FindRoyalFlush returns that entity, which is used to
determine the winner if a hand contains a royal flush.

We need to verify the :cards, attributes on the RoyalFlush entity
because we cannot perform comparisons during a tie without properly
saving the Card objects provided when a royal flush is identified.

We need to verify that FindRoyalFlush only returns a RoyalFlush
entity when a royal flush is found because misidentifying a poker
hand as a RoyalFlush would result in inaccurate identification of a
winner.