Spec for Pair
We need to verify that any instance of a Pair entity must have the
correct attributes because a different strength value for example would
change how a pair hand compares to others.

We need to verify that ties are resolved correctly so that the proper
winner is reported each time.