Spec for TwoPair
We need to verify that any instance of a TwoPair entity has the
correct attributes because a different strength value for example would
change how a two pair hand compares to others.

We need to verify that ties are resolved correctly so that the proper
winner is reported each time.