Spec for FullHouse
We need to verify that any instance of a FullHouse entity must have
the correct attributes because a different strength value for example
would change how a FullHouse hand compares to others.

We need to verify that ties are resolved correctly so that the proper
winner is reported each time.