Add wheel straight logic to FindStraightFlush
StraightFlush needs to have a high_card attribute like Straight to
maintain convention. A wheel straight flush needs to be able to be
recognized in FindStraightFlush for the proper high_card to be passed.
This lets us also change the logic for determining ties because we can
now compare high_card attributes.