Add BuildCards factory
We use a BuildCards factory so that the Deck class no longer knows how
to build cards itself. This makes testing easier and makes logical
sense.

For now we will assume that there are 13 ranks and 4 suits as is a
standard deck of cards that would be used in a game of poker. In the
future we could add a parameter to Deck that allows us to define what
a deck, for instance if we wanted suits to be emojis or if we wanted 20
ranks instead of 13.