Add Spec for the Deck class and BuildCards factory
We test that a standard deck is being created because the rest of our
program is built to play poker.

We test that Card objects are being properly removed from the deck
because we need to be able to form poker hands of a certain size.

For BuildCards we test that we are receiving Card objects. My reason
for only testing this is that we don't care what ranks and suits are,
we only care that we are getting Card objects from BuildCards. We can
make the ranks and suits whatever want, we just happen to be playing
poker and using a standard deck.