Add Spec for ClassifyHand
We test that the proper entity is returned when given a three of a kind
because the Class would be required to skip over nil entries in the
found_hands array and classify the best poker hand. By testing a poker
hand in the middle, we show that this is happening.